+++ src/HotelPosSystem/BookingForm.cs
@@ -1,4 +1,6 @@
-﻿using HotelPosSystem.Entities;
+﻿using System.ComponentModel;
+
+using HotelPosSystem.Entities;
 
 using Microsoft.EntityFrameworkCore;
 
@@ -7,9 +9,12 @@ namespace HotelPosSystem {
         private readonly BookingList _bookingList;
 
         private ComboBox? _customerDropdown;
+        private Label? _customerErrorLabel;
         private DateTimePicker? _startDatePicker;
         private DateTimePicker? _endDatePicker;
+        private Label? _dateErrorLabel;
         private ComboBox? _roomDropdown;
+        private Label? _roomErrorLabel;
         private TextBox? _commentTextBox;
         private CheckBox? _paidForCheckBox;
         private CheckBox? _checkedInCheckBox;
@@ -47,7 +52,10 @@ namespace HotelPosSystem {
             Customer[] customers = databaseContext.Customers
                 .OrderBy(customer => customer.FullName)
                 .ToArray();
-            (_customerDropdown, _) = ControlUtilities.AddComboBoxWithLabel(formPanel, "customer", customers, width, "customerLabel", "Customer:");
+            (_customerDropdown, _) = ControlUtilities.AddComboBoxWithLabel(formPanel, "customer", customers, "Select a customer", width, "customerLabel", "Customer:");
+            _customerDropdown.Validating += (object? sender, CancelEventArgs eventArgs) => ValidateCustomer();
+            _customerErrorLabel = ControlUtilities.AddLabel(formPanel, "customerError", string.Empty);
+            _customerErrorLabel.ForeColor = Color.Red;
 
             FlowLayoutPanel dateContainer = new() {
                 FlowDirection = FlowDirection.LeftToRight,
@@ -60,12 +68,19 @@ namespace HotelPosSystem {
             (_endDatePicker, _) = ControlUtilities.AddDatePickerWithLabel(dateContainer, "endDate", DateTime.Now, "endDateLabel", "End date:");
             _startDatePicker.Margin = new Padding(_startDatePicker.Margin.Left, _startDatePicker.Margin.Top, right: MainForm.MarginSize, _startDatePicker.Margin.Bottom);
             _endDatePicker.Margin = new Padding(left: 0, _endDatePicker.Margin.Top, _endDatePicker.Margin.Right, _endDatePicker.Margin.Bottom);
+            _startDatePicker.Validating += (object? sender, CancelEventArgs eventArgs) => ValidateDates();
+            _endDatePicker.Validating += (object? sender, CancelEventArgs eventArgs) => ValidateDates();
+            _dateErrorLabel = ControlUtilities.AddLabel(formPanel, "dateError", string.Empty);
+            _dateErrorLabel.ForeColor = Color.Red;
 
             Room[] rooms = databaseContext.Rooms
                 .Include(room => room.Type)
                 .OrderBy(room => room.Name)
                 .ToArray();
-            (_roomDropdown, _) = ControlUtilities.AddComboBoxWithLabel(formPanel, "room", rooms, width, "roomLabel", "Room:");
+            (_roomDropdown, _) = ControlUtilities.AddComboBoxWithLabel(formPanel, "room", rooms, "Select a room", width, "roomLabel", "Room:");
+            _roomDropdown.Validating += (object? sender, CancelEventArgs eventArgs) => ValidateRoom();
+            _roomErrorLabel = ControlUtilities.AddLabel(formPanel, "roomError", string.Empty);
+            _roomErrorLabel.ForeColor = Color.Red;
 
             const int checkboxContainerMaxHeight = 22;
             TableLayoutPanel checkBoxContainer = new() {
@@ -96,7 +111,8 @@ namespace HotelPosSystem {
             using HotelDbContext databaseContext = new();
 
             Customer? customer = _customerDropdown?.SelectedValue as Customer;
-            if (_startDatePicker is null || _endDatePicker is null) {
+            if (_startDatePicker is null || _endDatePicker is null
+                || _customerErrorLabel is null || _dateErrorLabel is null || _roomErrorLabel is null) {
                 throw new NullReferenceException();
             }
             DateOnly startDate = DateOnly.FromDateTime(_startDatePicker.Value);
@@ -106,8 +122,12 @@ namespace HotelPosSystem {
             bool? paidFor = _paidForCheckBox?.Checked;
             bool? checkedIn = _checkedInCheckBox?.Checked;
 
-            if (customer is null || room is null || paidFor is null || checkedIn is null) {
+            if (paidFor is null || checkedIn is null) {
                 throw new NullReferenceException();
+            }
+
+            if (!ValidateFormData()) {
+                return;
             }
 
             Booking booking = new() {
@@ -151,6 +171,45 @@ namespace HotelPosSystem {
 
             _paidForCheckBox.Checked = false;
             _checkedInCheckBox.Checked = false;
+        }
+
+        private bool ValidateFormData() {
+            return ValidateCustomer() & ValidateDates() & ValidateRoom();
+        }
+
+        private bool ValidateCustomer() {
+            Customer? customer = _customerDropdown?.SelectedValue as Customer;
+            if (customer is null) {
+                _customerErrorLabel.Text = "Please select a customer";
+                return false;
+            }
+
+            _customerErrorLabel.Text = string.Empty;
+            return true;
+        }
+
+        private bool ValidateDates() {
+            DateOnly startDate = DateOnly.FromDateTime(_startDatePicker.Value);
+            DateOnly endDate = DateOnly.FromDateTime(_endDatePicker.Value);
+
+            if (endDate < startDate) {
+                _dateErrorLabel.Text = "End date is before start date";
+                return false;
+            }
+
+            _dateErrorLabel.Text = string.Empty;
+            return true;
+        }
+
+        private bool ValidateRoom() {
+            Room? room = _roomDropdown?.SelectedValue as Room;
+            if (room is null) {
+                _roomErrorLabel.Text = "Please select a room";
+                return false;
+            }
+
+            _roomErrorLabel.Text = string.Empty;
+            return true;
         }
     }
 }