+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- Booking form validation
+
 ### Changed
 
 - Layout improvements
+- No customer/room is selected by default in booking form
 
 ## 0.3.0 - 2024-01-18